Getting lsblk kname in the cases where the model name isn't available
Summary:
lsblk command does not return model name as output in some specific cases, specially when you have a micro SD in your computer (perhaps it should happen with a normal SD card as well, but I haven't tested it).

So I included a verification in SfdiskBackend::scanDevice to check for kname in lsblk and use it as the name in these cases.

Reviewers: stikonas

Reviewed By: stikonas

Tags: #kde_partition_manager

Differential Revision: https://phabricator.kde.org/D16577