fs: Add new member to FileSystem class for filesystem-specific features
In order to be able to create filesystems with specific features enabled
(or not), we need to store these features, as a QStringList variable,
inside the FileSystem private structure.

This commit adds the corresponding member variable, changes the
contructor to include a new `features` parameter, and changes every
filesystem constructors in order to take this new variable into
account.

Future commits will implement the createWithFeatures() function for
filesystems which support it.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>